Not seeding if environment is test

+++ server.js
@@ -13,9 +13,11 @@ var express = require('express'),
 
 Associations.associate();
 
-db.sync({force: true}).then(function(){
-    seeder.seed();
-});
+if(process.env.NODE_ENV != 'test'){
+    db.sync({force: true}).then(function(){
+        seeder.seed();
+    });
+}
 
 app.use(function(req, res, next) {
     res.header('Access-Control-Allow-Origin', '*');
